+++ io/CHANGES.md
@@ -1,4 +1,6 @@
 # unreleased 0.4.2
+## Add
+- add `AsyncIo` impl to `&TcpStream` and `&UnixStream`
 ## Fix
 - relax trait bound of `io_uring::write_all`